pyproject.toml
[project]
name = "bug-detective"
version = "0.1.0"
description = "AI Bug Detective — Multi-Agent Error Analysis System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "litellm>=1.55.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "langfuse>=2.50.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "pytest-asyncio>=1.3.0",
    "numpy>=1.24.0",
    "boto3>=1.42.63",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]
semantic-search = [
    "faiss-cpu>=1.8.0",
]
semantic = [
    "faiss-cpu>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
